+++ python/src/init.rs
@@ -3,7 +3,7 @@ use std::sync::Arc;
 
 use murr::api::MurrHttpService;
 use murr::conf::{BackendConfig, Config, StorageConfig};
-use murr::io4::store::rocksdb::plain::PlainConfig;
+use murr::io::store::rocksdb::plain::PlainConfig;
 use murr::service::MurrService;
 
 pub fn build_config(cache_dir: String, http_port: Option<u16>) -> Config {
