--- src/io/info.rs
@@ -1,16 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-use crate::core::TableSchema;
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TableInfo {
-    pub schema: TableSchema,
-    pub segments: Vec<SegmentInfo>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SegmentInfo {
-    pub id: u32,
-    pub size_bytes: u32,
-    pub num_values: u32,
-}
