+++ src/io/store/snapshot.rs
