--- src/io4/mod.rs
@@ -1,6 +1,0 @@
-pub mod column;
-pub mod fs;
-pub mod row;
-pub mod schema;
-pub mod store;
-pub mod table;
