+++ src/lib.rs
@@ -2,7 +2,6 @@ pub mod api;
 pub mod conf;
 pub mod core;
 pub mod io;
-pub mod io4;
 pub mod service;
 pub mod util;
